style(esl-utils): split formatting utils to separate files
